+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 *   Set `mode` for ESLint JSDoc linter to `typescript`.
 *   Improved ESLint linting: now only files in the project dir are linted. This ensures that symlinked dev packages are
     not linted. (`/node_modules/` etc are still excluded, of course.)
+*   (feature) Add option to polyfill core node packages in webpack.
 
 
 8.1.0